Fix ReadRecord getRecords() implementation
Since a MusicRecord object was serialized into the .ser file in Part I, a MusicRecord object needs to be read when deserializing the file, not primitive data types.

A readme was added to help with compilation and running part II.

All copies of mySongs.ser have been deleted. To run part II, part I needs to first be run to create a mySongs.ser file. Then in the main method of ReadRecord.java, edit the path to the getRecords() filename to find the newly created mySongs.ser.

Signed-off-by: Thomas Hewton-Waters <thewton.waters@gmail.com>